[VisionGlass] Add voice search button to Phone UI
Added a new Voice Search button to the phone UI. It's only enabled
when the keyboard is shown, i.e., when input is required. Once
clicked the handling of the input is directly managed by the
KeyboardWidget which properly handles the different types of input
(like entering text in a Web page vs entering text in a UI widget).