Restore back action using B/Y buttons in non-WebXR mode
In a3f4cb4c we removed the possibility of going back with B/Y buttons
because that was interferring with potential action assignments of
those buttons inside WebXR experiences. We were exiting the WebXR
experience instead of executing the action that was set by authors.

However it turns out that these buttons are heavily used by users
in non-immersive sessions as they're very convenient as you don't
really have to hold both controllers to have full control of your
2D navigation.

Fixes #1266