Allow asking the soft keyboard from the out side of the window widget
This patch fixes the issue that Chromium doesn't type nothing in
the form. The problem was that the focused view notified from the
system is not involved in the window widget which the keyboard
widget is attached to. Since Chromium has a content view outside of
the windows widget's layer, the focus view could be different with
the window widget.

This patch allows asking the soft keyboard from the out side of
the window widget.