[VisionGlass] Retry the USB permission request in case of error
In order to show contents in the glasses Wolvic needs to have USB
permissions. The system asks by itself whenever the VisionGlass API
is used. However there is no way for us to know whether that permission
was requested or not, so we have to ask for it anyway in case we don't
have permissions.

If the user dismisses the two requests then we show an alert dialog and
exit the app.