[VisionGlass] Delay IMU initialization until presentation is ready
We were initializing the glasses IMU very early in the initialization
process. However we don't have to do that until the presentation is
ready because we wouldn't use the IMU data for anything.

This allows us to remove the mWasIMUStarted field. The checks were
it was used were replaced by a null check of the active presentation.