[VisionGlass] Call updateDisplays only for presentation displays
DisplayManager reports changes in all types of displays but we are
only interested in the presentation displays as we call updateDisplays()
in the callbacks.

Now that the updateDisplays calls are confined to this use case,
we know for sure that if there is no display available then it's
because the user refused to grant presentation permissions. In that
case show an error dialog and exit.