[OpenXR] Enable hand tracking motion range
The hand tracking motion range extensions allow applications to
specify the motion range of the user's hands, so that it could take
better decisions when computing the hand joints poses. In our case
we'll be assuming that the user is not handling anything in their
hands (otherwise it'd be using the controllers). Consequently we
ask the runtime to use the hands full motion range. This allows
the runtime to detect a closed fist for example.