[VisionGlass] Recenter view with first valid glasses pose
In VisionGlass the view is not properly shown "in front of" the user
in cases in which the user is moving their head while Wolvic is
starting. The reasons are multiple:
1. we were incorrectly getting identity quaternions sometimes
in positions where the glasses orientation was not precisely that
2. the recenter caused by setting the head lock state was not
issued on time, so it was not doing anything in the practice

From now on we don't render any frame until a valid head pose is
received and also we do perform a full widget recenter after
the first drawn frame.