[Vision Glass] Dynamic realign button
Add a button to realign the controller and the view.

This button displays a dynamic view of the controller's position
in order to better communicate its purpose.

The circle in the button tracks the controller position,
tapping the button will reset that position and place
the circle in the middle again.
This functionality is not implemented in this PR.