[VisionGlass] Store the orientation quaternion instead of the matrix
For each orientation quaternion we get from the glasses we need to
convert it to a transformation matrix to be able to manipulate it
at different levels (VRB, browser engine, etc...). However we can
instead keep the quaternion as is and do only the transformation
on StartFrame(). This way we'll require less storage space and
also we wouldn't need to perform that computation for every single
value we get from sensors, just when we really need it for rendering.