[VisionGlass] Add controller calibration support
The VisionGlass system was designed as an external display for 3D
content more than a real VR device. This means that using the phone
as a controller is an ad-hoc solution we're using more than
something that was carefully designed before. The consequence is that
there are things that don't work properly out of the box. The most
important one is likely the orientation. Both the glasses and the
phone use different coordinate systems to compute the orientation
so it isn't possible to automatically alineate them.

That's why we need the user assistance to calibrate the orientations.
This PR provides an API that can be called from UI to calibrate
the glasses and the phone orientations. The code will asume that
both are pointing to the same direction and then it will
perform adjustments for the incoming orientation data from devices.