Make some audio streaming services work by faking UA agents
We're adding support for 3 different audio streaming services that
were not working properly. Two of them were offering us their mobile
versions and the UI was totally broken for VR devices. Setting
desktop mode automatically for those sites make them work. The
sites were soundcloud.com and zingmp3.vn

For the other site that was not enough and we had to completely
replace the UA to the Firefox Desktop one. This fixed nhaccuatui.com.