[Chromium][HVR] Fix aim poses Chromium
Chromium expects aim poses to be based on grip poses. That's usually
achieved by retrieving the aim pose using the grip space as base space.
However that is not supported by the buggy HVR OpenXR runtime that
retrieves invalid poses in that case. That's why we need to workaround
that in Wolvic.

Basically we keep retrieving the aim pose as we used to but instead
of using it we generate our own grip pose with a simple transformation,
moving the grip pose up in the Y axis and slightly rotating on the X
axis.