[VisionGlass] Add scroll gesture to the phone touchpad
From now on users can scroll web pages by scrolling on the touchpad
of the phone UI. This involved a set of changes:

1. Migrate from onTouchEvent() to GestureManager.onTouchEvent().
Instead of handling the low level UP/DOWN events the code will
deal with onScroll, onSingleTap events recognized by the
GestureManager. Due to this change scrolling via tap and tilt
the controller (phone) is no longer possible.
2. Convert PlatformActivityPlugin into an abstract class. This
was done to allow some method implementations. We added a new
interface implemented by VRBrowserActivity to get scroll
notifications from the platform plugins.
3. Use the current event generation machinery to deliver
a scroll event to the currently hovered widget. This is generally
the web widget but it could be another one like the settings
dialog.