[Chromium] Remove Mozilla accounts from UI
In particular do not show neither the accounts settings dialog on the
first run and also hide the accounts setup button in the Settings.

We had to disable it for the beta release because the login page does
not load properly with a non-Gecko engine because the Mozilla AC try
to inject a Gecko extension to perform some actions. As we don't have
extensions support yet, the page fails to load.