[VisionGlass] Fix calibration computation
VisionGlass and the phone report different orientation values (in form
of quaternions) for the same physical orientation. That's why we need
to perform a manual calibration procedure to get aligned values.

Given two quaternions q1 (head) and q2 (controller), the correct way to
do that is the following:

1. q_rel = q1 * q2^-1 (being this the inverse of q2)
2. for each q3 future values from the controller, do
    q_calibrated = q_rel * q3

That way calibration procedure is correct.