[VisionGlass] Quaternion normalization & cleanups
Quaternion multiplication can return non normalized values and we need
the result to be normalized for the followup operations.

We're also removing a calibration status boolean attribute that is
not used at all.

Last but not least, the code that generates valid orientation data
for the head is now refactored in a private method as it's used
in multiple places.