Reorient immersive videos
This PR is a slight modification of the method `OpenXRLayer::Update`
to specify that the `XrPosef` parameter indicates the position and
orientation of the layer in the current space (i.e. "reorient").

This "reorient" pose information is used to change the orientation
of immersive videos through the `XrCompositionLayerEquirectKHR` layer
where these videos are drawn.

In the future, we may use this information to reorient other types
of layers around the user.