Add null checks before calling PlatformPlugin methods
In 6958a465 we added a listener interface to the platform plugins so
that the VRBrowserActivity could get events from them. The only
flavour implementing the platform plugin right now is the VisionGlass
one. That's why we still need to have null checks before any
plugin method call. The aforementioned commit was adding a couple
of unchecked calls and thus it was causing crashes for all non
VisionGlass flavours.