Rotate the splash logo with head's pose
Wolvic was only using the head's position to show the splash logo
completely ignoring the head's rotation. It was kind of assuming
that it wouldn't change since the program starts. But the reality
is that users move their heads while the splash is not shown and
thus the logo is sometimes not showing in front of the user.