Fix positioning and orientation of flat widgets with cylinder parents
The code that adjusts the position and orientation of a flat (quad)
widget that has a curved (cylinder) parent was not computing valid
positions and orientations. This was quite evident when checking,
for example, the position of the tray bar in multi-window (curved)
mode with non default window distances. We have replaced it by new
code that provides much better results.

Fixes #1372, #1373