Do not create intents for handling invalid URLs
The current code was trying to launch an intent to handle unrecognized
URLs by external apps. However it was doing it even for the cases
where the provided URL had no scheme, i.e., was not a valid URL.

It was assuming that if the web engine did not support the scheme then
it should be handled by an external app, but the problem was that
sometimes the String representing an URL was not a valid URL as it
didn't even have an scheme (for example a text coming from the URL
bar typed by users).

We were not seeing any pratical effect because isEngineSupportedScheme()
was returning true when receiving null schemes which is wrong. That
would be fixed in a follow up patch.