[Vision Glass] UI to realign controller on startup
Display a dialog informing the user that it is necessary to
point the phone straight ahead to set up the controller.
Clicking the realign button on it will re-calibrate the
controller and close the dialog.
A similar dialog is displayed on the VR environment.
The pointer and the window are re-centered at the same time.