Use the HandJoints gesture manager for detecting system gesture
Since Lynx 1.1.8 we can realy on the R_FB_hand_tracking_aim OpenX
Extension, so we can use the its specific Gesture Manager. It works
fine for the getting poses and managing pinching gestures, but the
XR_HAND_TRACKING_AIM_VALID_BIT_FB is not properly updated, so we
can not use it to detect system gestures.