Add Visitor to UrlUtils for WebEngine specific behaviours
The UrlUtils class is a set of class methods that deal with URL
manipulation. It was designed that way because there it was mainly
a collection of utility methods. However now that we have 2 engines
that approach is not valid anymore as some of the methods should
return different values depending on the engine.

That's why in order not to modify the existing behaviour we've
added a Visitor class that is passed to some of the methods to
provide WebEngine specifics.