Add support for additional URL schemes when typing in the URL bar
Users can type almost anything in the URL. Wolvic tries to
differentitate between valid web URLs that are passed directly
to the engine and any other text which is sent to the user
configured web search engine.

The current implementation was constrained to a very limitted
amount of schemes. By using the recently added UrlUtilsVisitor
we can expand the support not only to engine specific URLs
(like "chrome://" or "moz-extension://") but also to other
well supported web schemes that were previously filtered out
(like "data:").

This adds also a significant new amount of unit tests for
the UrlUtils::urlForText() method.

Fixes #1391