[Vision Glass] Beam extends all the way to the pointer
It is easy to get lost while using the Vision Glass because the
controllers are not drawn in the 3D world and the narrow
field of view means that the beam is not always visible to the user.

This change extends the beam so it reaches all the way to
the pointer, making the UI more understandable.