Clean-up of resizing logic
Clean-up of our resizing logic to make it more consistent and
to remove some edge cases. This logic is now mostly
encapsulated inside of WindowWidget.

Manual mode allows the window to be resized freely within the
allowed boundaries. The maximum width depends on the number
of active windows. The maximum height has been increased
slightly to accommodate vertical videos better.

WindowWidget.getSizeForScale() is used to resize the window
when using one of the preset sizes and when entering
fullscreen. It will keep the window size within the
boundaries while trying to preserve the current aspect ratio.