[ML2] Fix a crash on release builds
The recently added support for XR_ML_frame_end_info extension
inadvertedly introduced a crash in release builds. The problem is
that we're adding to the ".next" pointer of XrFrameEndInfo a
pointer to a struct that was defined inside an if block. This means
that the struct is freed when going out of scope. That is not
generally a problem on debug builds but it causes crashes in
release ones (and it's incorrect anyway).