Fix a potential crash on release when overlay extension is enabled
The overlay extension is enabled by passing a special struct
to the next pointer of the XrSessionCreateInfo struct. The problem
was that we were defining that special struct inside an if block.
This means that when the if block is gone so the variable and
thus that could cause crashes on release builds (also it's
incorrect anyway).