Remove the OculusVR backend
We've been using OpenXR for Meta devices for a while. Actually new apps
in the Meta Store are not allowed to use the OculusVR backend. We kept
it for a while as a reference for potentially missing features in the
OpenXR backend. The truth is that we haven't checked it for quite some
time now and it's actually becoming a maintenance burden. For historical
purpouses we have git.