[VisionGlass] Add position capabilities to the WebXR controller
The controller was not shown in WebXR sessions. It was actually not
possible to interact with any experience. The problem was that
we were not setting position capabilities to the controller and
thus we were not passing that information to the engine (even if
the position is emulated as the phone is a 3DoF device).
After this change the controller can be seen in WebXR.

Also we're setting the number of buttons in the controller to 1
to let experiences adapt to a controller with just one button.