[VisionGlass] Restore WebXR rendering
WebXR scenes were not rendered in the VisionGlass since 0f55393a6.
The regression was caused by an error specifying the FoV of the
immersive display. We were incorrectly using negative sizes for left
and bottom as they must be positive. The code is smart enough to
compute the FoV in the four directions using absolute figures.