[VisionGlass] Make seek backward/forward work with non-fullscreen media
The previous implementation was relying on the Media's seekTo() method
and, specifically for this case, on the getCurrentTime(). The problem
of the latter is that it depends on receiving position changed events
from the Web engine. In the case of Gecko, for whatever reason, those
events are only emitted when in fullscreen.

By switching to using seekForward()/seekBackward() we can keep the
functionality. The only significat change is that we cannot control
the amount of time skipped, which will be now 5s instead of the
previous 10s for backwards and 30s for forwards.