[Vision Glass] Fix theming issues
The main application (and therefore the UI for VR) uses FxR_Dark,
which fixes all the glitches with text colors, theming, etc.

The theme for the phone UI is applied with a LayoutInflater which
uses a different theme, which ensures that the correct values are
applied in a way that does not interfere with the VR interface.

Colors of the status and navigation bars are specified in FxR.Dark
to ensure that they are properly applied. This doesn't affect VR.