[VisionGlass] Make the sensor orientation transformation code self-explanatory
We need to transform the quaternion returned by the phone orientation
sensor in order to make it usable in Wolvic. So far it involves two
transformations:
1. changing the order of the components of the quaternion: we need
[x,y,z,w] but we get [w,x,y,z]
2. swap y and z axes: that's because the phone is meant to be used
in portrait mode not flat as a TV remote as we use it

This is done now in a single line when creating the quaternion but
that's pretty obscure. Better split it in the two operations
mentioned above even though it involves more but supercheap operations.