[VisionGlass] Remove grip capabilities from controller
We should not really enable that flag because our controller (the
mobile phone) is not able to provide a pose for the aim and for
the grip as both are the same in this case.

By disabling this we prevent Wolvic and the web engine from
computing something that does not exist.