[VisionGlass] Make 1€ filter more responsive to high speed movements
By increasing the beta parameter we make the filter more responsive
to high speed movements giving the feeling that the controller is
less laggy. We don't really need to filter those, the main issue
is having a shaky pointer when handing the phone more or less still.
As we don't change the fcmin, the filtering for small movements,
we are still removing the noise from the signal.