[Chromium] Fix crash in media permissions
Requesting an empty list of either video or audio media sources is
currently causing a crash because Wolvic expects aAudio or aVideo to be
null in this case but Chromium provides an empty list instead. This
patch makes the code more robust by supporting both an empty list and a
null object.