[VisionGlass] Fix controller gimbal lock
We're suffering from gimbal lock in the controller orientation (and
this in the pointer) which was pretty evident as soon as we turn
our head 90 degs on either direction. The consequence as with any
gimbal lock was that axes appeared to be swapped, like suddenly a
rotation in the z axis caused changes in the x axis. Also changes
in some axis were not causing any effect. Those are well known
results of having a gimbal lock.

In order to fix it we first had to fix the calibration. We now
store the head position when calibration happens. That corresponds
to the phone identity orientation (as we reset the IMU). This
means that if we revert that rotation (conjugate) on every frame,
we make the phone identity position equivalent (from the world
coordinates POV) to the stored head position during calibration.
Then we need to use the conjugate of the resulting quaternion.

There was a remaining thing to fix, we need to invert x and y axis
so that moving the phone to the right causes the pointer to move
to the right too. Same for y axis but with up/down directions.