[OpenXR] Make Wolvic run without active interaction profile
Interaction profiles define a collection of buttons and other input
sources (like poses) in a physical arrangement which allow apps to
set mappings between those input sources and actions.

For the case of hand tracking, we don't really need any interaction
profile yet as we're using extensions to gather hand data and
emulate a controller.

The current code assumed that we must always had an active
input mapping. We were workarounding it by using the default
Khronos interaction profile for devices without controllers.
But that is not really needed. The right fix is to allow the
code to run without any active interaction profile.