[OpenXR] Do not use input mappings to set the device type
Setting the correct device type is important among other things
to load the proper 3D models for controllers, set the right
input mappings or select the correct WebXR input profile.

We used to pick the device type from the mapping, but that is
not a good idea because in the future we could have input
mappings shared by different devices. Also there might be the
case that a device has no input mapping (like devices
without controllers).

For that reason it's better to store the device type that
we retrieve from the system and use it.