[OpenXR] Remove Khronos simple interaction profile
We do have interaction profiles for all the supported
devices with physical controllers. We don't really need
it for the hand tracking only devices which should
work fine after the previous work that allows Wolvic
to run without any active mapping.

Allowing multiple interaction profiles has some
drawbacks and thus it's better to remove it. For example
in the case of Meta devices, if the Khronos simple
interaction profile is enabled, then we start to
receive not valid hand tracking data (like all the
flags being zero).

Fixes #1430