Create the beam after creating controllers
Currently the beam for the controllers is created only after the
controllers are in a "ready" state. That can vary among backends
but for example in the case of OpenXR that happens only after
having a valid interaction profile for the controllers.

That makes sense for devices with physical controllers with
interaction profiles but does not apply to hand tracking only
devices that might not have interaction profiles.

In reality, to create the beam, we do only need the
ControllerDelegate objects to be created which normally happens
very early (at the DeviceDelegate creation/initialization). That's
why we split the beam initialization and controllers' 3D model
loading in two different callbacks. The 3D model will still happen
in the ready callback but the beam initialization is moved to
an early stage, to the create callback.

By proceeding this way we will always create the beam even
without interaction profiles, and for devices with both
physical controllers and hand tracking, the 3D model loading
will still work even if Wolvic is started in hand tracking mode.