Regression: hand interaction profile is not available
When we removed the KHR simple profile in c3b539a96 we also did
delete the code that was handling profiles with no specific
device. That was a mistake because that's precisely the case
of generic profiles like hand interaction.

Restored the code adding those profiles to the list of available
mappings.