[Vision Glass] Back button when VR is not active
Fix the issue that the Back button on the phone is not responsive
before we have entered VR mode.

The fix is to add a method onBackPressed() to PlatformActivityPlugin
which returns true if the platform activity has handled the Back press.

In the Vision Glass implementation, we use the default Back behaviour
if the user presses that button before we have started VR mode.