[Vision Glass] Voice button can be used for search and text input
Set up the connection to the Huawei Speech Recognition service.

The voice button on the phone UI will perform a search if the user
is not currently writing on a text field (and therefore the keyboard
is not visible).

When the keyboard is visible, the voice button will capture text which
will be entered into the text field that the user is editing.