[HVR] Move shared functionality to "hvrcommon"
A number of classes and layouts are shared by the builds for HVR
and Vision Glass, for example the ones that use Huawei services
and provide phone-specific UI elements (e.g. legal terms).

These sources will be located in the "hvrcommon" folder so it is
easier to share them between both builds.