[OpenXR] Do not process button states when not ready
Some extension provide the /ready_ext path that allows the runtime
to provide feedback about whether a specific action is ready to
be triggered or not. One example is the input/pinch/ext_ready path
from the hand interaction profile extension.

We were using the ready value to filter out events on trigger
buttons but we should really consider it for any of other inputs
(even if there is not such a thing in the current mappings)