[Chromium] Fix text input fields not getting keyboard focus
This sequence of events:
1. Goto Disney+ and click on the email text entry
2. Go home and click on Filmin link
3. Click on the email entry text

Step 3 was failing as it was not apparently possible to get the
keyboard focus (actually entering text was not possible but
it was possible to paste or delete text).

This patch fixes this issue by only adding the TabCompositorView
to the widget container. The code used to add the Tab's
ContentView too. I don't understand that code well enough, but
my guess is that there was a race between the two and sometimes
one was getting the focus and some other times the other. It's
still unclear to me why the above sequence was always failing.

The text input is fixed and apparently there are no regressions.

Fixes #1345