[OpenXR] Eye tracking support for navigation
This PR provides an implementation for eye tracking based navigation
àla Apple VisionPro (not as good yet but quite usable).

It's implemented using OpenXR's XR_EXT_eye_gaze_interaction extension.
This extension provides a path to retrieve gaze input from the runtime.
Wolvic uses that data as the aim. All the other input interactions
remain the same, meaning that users can use hand tracking or the
physical controllers to perform actions like clicks or scrolling.

Speaking about scrolling, some extra code had to be added to properly
handle scrolling. This is the rationale, when not using eye tracking
scrolling happens when the aim is moved (up/down) while pressing
the trigger button (or a pinch with hand tracking). However when eye
tracking is enabled we cannot do that, because we don't want to
scroll by moving our eyes but by moving either the hand or the
controller we are using for clicking/pinching. That's why some
code to detect scrolling (based on heuristics) was added for the
eye tracking case.

Eye tracking is a very sensitive technology from the privacy POV
as it can be used for fingerprinting for example. That's why Wolvic
asks for permissions before using it. In case users reject the
request the usual hand tracking or controllers input will be used
instead. For devices supporting eye tracking, a new radio button
group is now visible in the controller options in settings.

The eye tracking implementation supports devices using both
hand interaction profiles (like the MagicLeap2) or controller
emulation via hand tracking (like the Pico4E). This implementation
was tested on both devices, but it might work well in others
supporting the eye gaze extension.