Specify the controller Id when triggering haptic feedback
We are sending the haptic feedback to the controller that
got the focus due to the last click event. This causes that
hovering events generated by the other controller still
send the haptic feedback to the inactive controller.

This change defines a new argument in the triggerHapticFeedback
function to pass the Id associated to the controller that
generated the hovering event.