Migrate Wolvic to Gecko v128 ESR
The v128 release is just 1 month old and the nice thing is that
it's an extended support release.

This required just one API change, a new parameter was added to
onLocationChange().

Some crashes appeared due to incompatibilities with the version
of AndroidComponents (v121) that we're using which is too old
for Gecko v128. The required migration of AC will be done in
a folloup patch.