New Controller's setting to disable hand tracking
There maybe be cases where the user wants to prevent the hand tracking
to take the control when the controllers got untracked.

This change defines a new switch in the Controller's settings dialog to
to disable the hand trakckig capabilities. This setting has been defined
as non-persistent, so it will be reset just restarting Wolvic.