Bring back the title bar widget
Amazingly the title bar widget visibility has been broken for almost
two years without anyone realizing about it. I figured it out by chance
after doing some testing with the old FxR and then voilà the title bar
was there with the very cool play/pause media button.

The visibility was broken in 136950b0 when Kiosk mode was added to
Wolvic. The problem was that the visibility observer was returning
false whenever kiosk mode was disabled instead of doing it when
kiosk mode was enabled (as there are not multiple windows in kiosk).