Fix assertions logic
The Assertions.h file was using a Throw() inline method defined in
OpenXRHelpers.h in several macros. There were several problems with
that:
1. The macros would only work if the file including Assertions.h was
also including OpenXRHelpers.h. That's wrong, Assertions.h should not
depend on that
2. Assertions.h should not have any dependency with OpenXR headers

This was unnoticed because files using those macros were already
including OpenXRHelpers.h. As soon as I tried to use it in non
OpenXR files it started to generate compiler errors.