Add a version check activity
This new activity becomes the new MAIN activity, so it'll be launched
before anything else including the VRBrowserActivity. It'll perform
an OS version check. On success it'll invoke the VRBrowserActivity.
Should the test fail, a dialog with a text asking user to upgrade the
OS will be shown and Wolvic will refuse to start.

So far the test is only performed for Meta builds (because the Khronos
OpenXR loader is only available in v62+ VROS. Nevertheless it could be
easily extended to other flavours.

I've verified that it works as expected in: Pico4, Quest2, HVR,
VisionGlass, Lenovo VRX, LynxR1 and the NoAPI flavour.